+++ Sources/MapboxSearch/PublicAPI/MapboxSearchVersion.swift
@@ -1,2 +1,2 @@
 /// Mapbox Search SDK version variable
-public let mapboxSearchSDKVersion = "2.12.0-beta.2"
+public let mapboxSearchSDKVersion = "2.12.0-rc.1"