+++ Sources/MapboxSearch/PublicAPI/Search Results/ExternalRecordPlaceholder.swift
@@ -24,6 +24,8 @@ class ExternalRecordPlaceholder: SearchResultSuggestion, CoreResponseProvider {
 
     var categories: [String]?
 
+    var categoryIds: [String]?
+
     var suggestionType: SearchSuggestType = .POI
 
     var distance: CLLocationDistance?
@@ -48,6 +50,7 @@ class ExternalRecordPlaceholder: SearchResultSuggestion, CoreResponseProvider {
         self.distance = coreResult.distanceToProximity
         self.originalResponse = CoreSearchResultResponse(coreResult: coreResult, response: response)
         self.categories = coreResult.categories
+        self.categoryIds = coreResult.categoryIDs
         self.serverIndex = coreResult.serverIndex?.intValue
         self.estimatedTime = coreResult.estimatedTime
         self.descriptionText = coreResult.addresses?.first.map(Address.init)?.formattedAddress(style: .medium)